backend/requirements.txt
# Core FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
python-multipart==0.0.6

# Claude API
anthropic>=0.39.0

# Database - PostgreSQL with async support
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Task Queue - Celery with Redis
celery[redis]==5.3.6
redis==5.0.1

# HTTP Client
httpx==0.26.0

# Web Scraping
beautifulsoup4==4.12.3

# Data Processing
fuzzywuzzy==0.18.0
python-Levenshtein==0.25.0

# Address Parsing (optional - install if needed)
# usaddress==0.5.10
# postal==1.1.10

# AWS S3 (optional - for image caching)
boto3==1.34.34

# Monitoring
prometheus-client==0.19.0

# Apify (optional - for Apify SDK)
# apify-client==1.6.0

# Auth
PyJWT[crypto]==2.9.0
supabase==2.11.0

# Payments
stripe==11.4.1

# Email
resend==2.5.0

# Development
pytest==8.0.0
pytest-asyncio==0.23.3
